Functioning cleaned tuning
Model theoretically can produce outputs, but its taken over 30 min to produce one output
