--- Test_Trained_Model.py
@@ -1,42 +1,0 @@
-# Databricks notebook source
-import logging
-import pandas as pd
-
-from transformers import AutoTokenizer, AutoModelForCausalLM, TrainingArguments, GPTQConfig, GenerationConfig
-# from datasets import Dataset
-# from sklearn.model_selection import train_test_split
-
-
-from time import perf_counter
-from rich import print
-
-import pprint
-import json
-from timeit import default_timer as timer
-
-from langchain.llms import LlamaCpp
-from langchain.callbacks.manager import CallbackManager
-from langchain.callbacks.streaming_stdout import StreamingStdOutCallbackHandler
-
-from peft import LoraConfig, get_peft_model, prepare_model_for_kbit_training, AutoPeftModelForCausalLM
-
-from trl import SFTTrainer
-
-from utils import fix_key_names, input_preprocessing
-
-from model_setup import ModelSetup
-from trainer_setup import ModelTrainer
-
-import mlflow
-import mlflow.sklearn
-from mlflow.models import infer_signature
-
-from datetime import datetime
-
-from defaults import get_default_LORA_config, get_default_training_args
-from delta.tables import DeltaTable
-from pyspark.sql.functions import col
-
-# COMMAND ----------
-
-
