--- predicted-cleaned.jsonl
@@ -1,14 +1,0 @@
-[
-    {
-        "input": "DUMMY DATA POINT 1",
-        "output": "{'DUMMY OUTPUT 2': '\\n                   {\\n    \"KEY1\": \"7\",\\n    \"KEY2\": \"bye\"\\n}\\n                   '}",
-        "preprocessed_input": "\n            Populate a JSON in the JSON_SCHEMA format from the provided TEXT_DATA.\n            ---\n            JSON_SCHEMA: \n{\n    \"KEY1\": \"int\",\n    \"KEY2\": \"str\"\n}\n\n            ---\n            TEXT_DATA: DUMMY DATA POINT 1\n        ",
-        "text": "\n    INPUT:\n    \n            Populate a JSON in the JSON_SCHEMA format from the provided TEXT_DATA.\n            ---\n            JSON_SCHEMA: \n{\n    \"KEY1\": \"int\",\n    \"KEY2\": \"str\"\n}\n\n            ---\n            TEXT_DATA: DUMMY DATA POINT 1\n        \n    \n    ----------------\n    \n    OUTPUT:\n    {'DUMMY OUTPUT 2': '\\n                   {\\n    \"KEY1\": \"7\",\\n    \"KEY2\": \"bye\"\\n}\\n                   '}\n    "
-    },
-    {
-        "input": "DUMMY DATA POINT 2",
-        "output": "{'DUMMY OUTPUT 2': '\\n                   {\\n    \"KEY1\": \"70000\",\\n    \"KEY2\": \"hello\"\\n}\\n                   '}",
-        "preprocessed_input": "\n            Populate a JSON in the JSON_SCHEMA format from the provided TEXT_DATA.\n            ---\n            JSON_SCHEMA: \n{\n    \"KEY1\": \"int\",\n    \"KEY2\": \"str\"\n}\n\n            ---\n            TEXT_DATA: DUMMY DATA POINT 2\n        ",
-        "text": "\n    INPUT:\n    \n            Populate a JSON in the JSON_SCHEMA format from the provided TEXT_DATA.\n            ---\n            JSON_SCHEMA: \n{\n    \"KEY1\": \"int\",\n    \"KEY2\": \"str\"\n}\n\n            ---\n            TEXT_DATA: DUMMY DATA POINT 2\n        \n    \n    ----------------\n    \n    OUTPUT:\n    {'DUMMY OUTPUT 2': '\\n                   {\\n    \"KEY1\": \"70000\",\\n    \"KEY2\": \"hello\"\\n}\\n                   '}\n    "
-    }
-]
