--- tuned_models/checkpoint-30/trainer_state.json
@@ -1,201 +1,0 @@
-{
-  "best_metric": null,
-  "best_model_checkpoint": null,
-  "epoch": 30.0,
-  "eval_steps": 500,
-  "global_step": 30,
-  "is_hyper_param_search": false,
-  "is_local_process_zero": true,
-  "is_world_process_zero": true,
-  "log_history": [
-    {
-      "epoch": 1.0,
-      "learning_rate": 1.9333333333333333e-05,
-      "loss": 2.657,
-      "step": 1
-    },
-    {
-      "epoch": 2.0,
-      "learning_rate": 1.866666666666667e-05,
-      "loss": 2.5372,
-      "step": 2
-    },
-    {
-      "epoch": 3.0,
-      "learning_rate": 1.8e-05,
-      "loss": 2.4094,
-      "step": 3
-    },
-    {
-      "epoch": 4.0,
-      "learning_rate": 1.7333333333333336e-05,
-      "loss": 2.2924,
-      "step": 4
-    },
-    {
-      "epoch": 5.0,
-      "learning_rate": 1.6666666666666667e-05,
-      "loss": 2.1907,
-      "step": 5
-    },
-    {
-      "epoch": 6.0,
-      "learning_rate": 1.6000000000000003e-05,
-      "loss": 2.1026,
-      "step": 6
-    },
-    {
-      "epoch": 7.0,
-      "learning_rate": 1.5333333333333334e-05,
-      "loss": 2.0144,
-      "step": 7
-    },
-    {
-      "epoch": 8.0,
-      "learning_rate": 1.4666666666666666e-05,
-      "loss": 1.9377,
-      "step": 8
-    },
-    {
-      "epoch": 9.0,
-      "learning_rate": 1.4e-05,
-      "loss": 1.8639,
-      "step": 9
-    },
-    {
-      "epoch": 10.0,
-      "learning_rate": 1.3333333333333333e-05,
-      "loss": 1.7917,
-      "step": 10
-    },
-    {
-      "epoch": 11.0,
-      "learning_rate": 1.2666666666666667e-05,
-      "loss": 1.7247,
-      "step": 11
-    },
-    {
-      "epoch": 12.0,
-      "learning_rate": 1.2e-05,
-      "loss": 1.663,
-      "step": 12
-    },
-    {
-      "epoch": 13.0,
-      "learning_rate": 1.1333333333333334e-05,
-      "loss": 1.6016,
-      "step": 13
-    },
-    {
-      "epoch": 14.0,
-      "learning_rate": 1.0666666666666667e-05,
-      "loss": 1.5459,
-      "step": 14
-    },
-    {
-      "epoch": 15.0,
-      "learning_rate": 1e-05,
-      "loss": 1.4922,
-      "step": 15
-    },
-    {
-      "epoch": 16.0,
-      "learning_rate": 9.333333333333334e-06,
-      "loss": 1.4428,
-      "step": 16
-    },
-    {
-      "epoch": 17.0,
-      "learning_rate": 8.666666666666668e-06,
-      "loss": 1.3978,
-      "step": 17
-    },
-    {
-      "epoch": 18.0,
-      "learning_rate": 8.000000000000001e-06,
-      "loss": 1.3538,
-      "step": 18
-    },
-    {
-      "epoch": 19.0,
-      "learning_rate": 7.333333333333333e-06,
-      "loss": 1.31,
-      "step": 19
-    },
-    {
-      "epoch": 20.0,
-      "learning_rate": 6.666666666666667e-06,
-      "loss": 1.2706,
-      "step": 20
-    },
-    {
-      "epoch": 21.0,
-      "learning_rate": 6e-06,
-      "loss": 1.2368,
-      "step": 21
-    },
-    {
-      "epoch": 22.0,
-      "learning_rate": 5.333333333333334e-06,
-      "loss": 1.2066,
-      "step": 22
-    },
-    {
-      "epoch": 23.0,
-      "learning_rate": 4.666666666666667e-06,
-      "loss": 1.1758,
-      "step": 23
-    },
-    {
-      "epoch": 24.0,
-      "learning_rate": 4.000000000000001e-06,
-      "loss": 1.146,
-      "step": 24
-    },
-    {
-      "epoch": 25.0,
-      "learning_rate": 3.3333333333333333e-06,
-      "loss": 1.1209,
-      "step": 25
-    },
-    {
-      "epoch": 26.0,
-      "learning_rate": 2.666666666666667e-06,
-      "loss": 1.1107,
-      "step": 26
-    },
-    {
-      "epoch": 27.0,
-      "learning_rate": 2.0000000000000003e-06,
-      "loss": 1.0876,
-      "step": 27
-    },
-    {
-      "epoch": 28.0,
-      "learning_rate": 1.3333333333333334e-06,
-      "loss": 1.0784,
-      "step": 28
-    },
-    {
-      "epoch": 29.0,
-      "learning_rate": 6.666666666666667e-07,
-      "loss": 1.0634,
-      "step": 29
-    },
-    {
-      "epoch": 30.0,
-      "learning_rate": 0.0,
-      "loss": 1.0594,
-      "step": 30
-    }
-  ],
-  "logging_steps": 1,
-  "max_steps": 30,
-  "num_input_tokens_seen": 0,
-  "num_train_epochs": 30,
-  "save_steps": 500,
-  "total_flos": 192109404487680.0,
-  "train_batch_size": 1,
-  "trial_name": null,
-  "trial_params": null
-}
