--- tuned_models/model3/config.json
@@ -1,37 +1,0 @@
-{
-  "_name_or_path": "/dolly_v2_3b/model",
-  "architectures": [
-    "GPTNeoXForCausalLM"
-  ],
-  "attention_bias": true,
-  "attention_dropout": 0.0,
-  "bos_token_id": 0,
-  "classifier_dropout": 0.1,
-  "custom_pipelines": {
-    "text-generation": {
-      "impl": "instruct_pipeline.InstructionTextGenerationPipeline",
-      "pt": "AutoModelForCausalLM",
-      "tf": "TFAutoModelForCausalLM"
-    }
-  },
-  "eos_token_id": 0,
-  "hidden_act": "gelu",
-  "hidden_dropout": 0.0,
-  "hidden_size": 2560,
-  "initializer_range": 0.02,
-  "intermediate_size": 10240,
-  "layer_norm_eps": 1e-05,
-  "max_position_embeddings": 2048,
-  "model_type": "gpt_neox",
-  "num_attention_heads": 32,
-  "num_hidden_layers": 32,
-  "rope_scaling": null,
-  "rotary_emb_base": 10000,
-  "rotary_pct": 0.25,
-  "tie_word_embeddings": false,
-  "torch_dtype": "float32",
-  "transformers_version": "4.37.1",
-  "use_cache": true,
-  "use_parallel_residual": true,
-  "vocab_size": 50280
-}
