--- tuned_models/model3/generation_config.json
@@ -1,6 +1,0 @@
-{
-  "_from_model_config": true,
-  "bos_token_id": 0,
-  "eos_token_id": 0,
-  "transformers_version": "4.37.1"
-}
