--- tuned_models/testing_safe_serial_model/adapter_config.json
@@ -1,29 +1,0 @@
-{
-  "alpha_pattern": {},
-  "auto_mapping": null,
-  "base_model_name_or_path": "/dolly_v2_3b/model",
-  "bias": "none",
-  "fan_in_fan_out": false,
-  "inference_mode": true,
-  "init_lora_weights": true,
-  "layers_pattern": null,
-  "layers_to_transform": null,
-  "loftq_config": {},
-  "lora_alpha": 32,
-  "lora_dropout": 0.05,
-  "megatron_config": null,
-  "megatron_core": "megatron.core",
-  "modules_to_save": null,
-  "peft_type": "LORA",
-  "r": 16,
-  "rank_pattern": {},
-  "revision": null,
-  "target_modules": [
-    "dense_h_to_4h",
-    "query_key_value",
-    "dense",
-    "embed_out",
-    "dense_4h_to_h"
-  ],
-  "task_type": "FEATURE_EXTRACTION"
-}
